cloud-run-ocr/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-cloud-firestore==2.13.1
google-cloud-core==2.4.1
requests==2.31.0
gunicorn==21.2.0
